Various fixes
* Uncompressed size was not being shipped
* Limits were parsed in 32-bits, which made some of them overflow
* Paths in runner `.sha1` files were wrong
* Explicitly fail with `.cat` files, since they are still not supported.

+++ common/context.go
@@ -31,7 +31,7 @@ type RunnerConfig struct {
 	RuntimePath         string
 	CompileTimeLimit    int
 	CompileOutputLimit  int
-	JavaVmEstimatedSize int
+	JavaVmEstimatedSize int64
 	PreserveFiles       bool
 }
